hw9 6try

+++ src/routes/auth-router.ts
@@ -136,15 +136,17 @@ authRouter.post('/refresh-token', async (req: Request, res: Response) => {
 authRouter.post('/logout', async (req: Request, res: Response) => {
     const refreshToken = req.cookies.refreshToken
     if (!refreshToken) return res.sendStatus(401)
-
     const expiredOrNot = await jwtService.getPayloadByToken(refreshToken)
     if (!expiredOrNot) return res.sendStatus(401)
     const {userId, deviceId} = expiredOrNot;
-    //const isExpiredToken = await jwtTokensService.isExpiredToken(refreshToken)
-    //if (isExpiredToken) return res.sendStatus(401)// check need i this verification or this redundant
+    const getSessionByUserIdAndDeviceId = await sessionService.getSessionByUserIdAndDeviceId(userId, deviceId)
+    if (!getSessionByUserIdAndDeviceId) return res.sendStatus(401)
+    if (expiredOrNot.iat !== getSessionByUserIdAndDeviceId.issuedAt)return  res.sendStatus(401)
+        //const isExpiredToken = await jwtTokensService.isExpiredToken(refreshToken)
+        //if (isExpiredToken) return res.sendStatus(401)// check need i this verification or this redundant
 
-    //await jwtTokensService.expiredTokens(refreshToken)
-    const result = await sessionService.deleteSessionDevice(userId, deviceId)
+        //await jwtTokensService.expiredTokens(refreshToken)
+        const result = await sessionService.deleteSessionDevice(userId, deviceId)
     if (result) return res.sendStatus(401)
     return res.clearCookie('refreshToken').sendStatus(204)
 })
